+++ radp-archetypes/scaffold-simple/src/main/resources/archetype-resources/src/main/java/app/config/ThreadPoolConfigProperties.java
@@ -5,7 +5,7 @@ package ${package}.app.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import ${package}.spring.framework.bootstrap.constant.Globals;
+import space.x9x.radp.spring.framework.bootstrap.constant.Globals;
 
 /**
  * @author x9x