+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [ ] fix archetype GitLab CI/CD yml
 - [x] fix archetype auto-release profile not work
+- [x] fix archetype import error
 
 ## 0.10